+++ src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import Header from './components/header';
+import Intro from './components/Intro';
 import { Card } from './components/Card';
 import Footer from './components/Footer';
 import TitleSection from './components/Title';
@@ -8,48 +8,13 @@ import TitleSection from './components/Title';
 
 function App() {
 
-  const features = [
-    {
-      // src: ,
-      title: "Chatbot",
-      description:
-        "Yea our chatbot is pretty cool",
-    },
-    {
-      // src: ,
-      title: "Data Visualization",
-      description:
-        "some graphs and forecasting stuff",
-    },
-    {
-      // src: ,
-      title: "News",
-      description: "Follow the most recent stuff for your stocks",
-    },
-    {
-      // src: ,
-      title: "Blah blah",
-      description:
-        "Another feature maybe",
-    },
-  ];
-
+  
 
   return (
-    <div className="home">
-      <Header />
-      <TitleSection />
-      <div className='features-list-container'>
-        {
-          features.map((feature, index) => {
-            return (
-              <Card key={index} title={feature.title} imgUrl={feature.src}>{<p>{feature.description}</p>}</Card>
-            )
-          })
-        }
+    <div>
+      <Intro />
       </div>
-      <Footer />
-    </div>
+      
   );
 }
