--- src/components/Footer.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-const Footer = () => {
-  return (
-    <div className="footer-container">
-      <div className="footer-copyright">
-        <p>Copyright 2022 InvestX. All Rights Reserved</p>
-      </div>
-    </div>
-  );
-};
-
-export default Footer;
