--- src/components/Title.jsx
@@ -1,24 +1,0 @@
-import React from "react";
-
-
-const TitleSection = () => {
-  return (
-    <div className="title-section-container">
-        <div className="title-info-text">
-          <h1>
-            The <span className="highlighted">Financial</span> Tool to Succeed
-          </h1>
-          <p className="title-info-description">
-            Free to use financial to use to keep you track with investments!
-          </p>
-          <div className="search-container">
-        
-          </div>
-          
-        </div>
-    
-    </div>
-  );
-};
-
-export default TitleSection;
